reject button made run

+++ endpoints/friends.js
@@ -4,6 +4,8 @@ const cors= require('cors')
 const jwt = require('jsonwebtoken')
 const User = require('../models/user')
 
+const Friend = require('../models/friend')
+
 const friendsRouter = express.Router();
 
 mongoose.connect(process.env.MONGO_URI, {
@@ -19,16 +21,7 @@ db.on('error', console.error.bind(console, 'MongoDB connection error:'));
 
 
 
-const friendSchema = new mongoose.Schema ({
 
-user1: {type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true},
-user2: {type: mongoose.Schema.Types.ObjectId, ref: 'User', required: true},
-areFriends : {type: Boolean, default: false},
-request : {type: Boolean, default: false}
-
-
-});
-const Friend = mongoose.model('Friend',friendSchema); 
 
 
 
@@ -37,7 +30,7 @@ friendsRouter.use(cors())
 
 friendsRouter.post('/request',async(req,res)=>{
 
-    console.log('request',req.body);
+    
 
     const from= jwt.verify(req.body.token, "THISISMYSECRETKEY");
 
@@ -66,7 +59,7 @@ friendsRouter.get('/getrequest',async(req,res)=>{
 
     let allRequest = await Friend.find({$or :[{user2:from.userid, request:true}]});
 
-    console.log('allrequest',allRequest)
+   
       
     // let allRequest2 = await Friend.find({user1:from.userid,request: true});
     // console.log('allrequest2',allRequest2)
@@ -138,20 +131,20 @@ friendsRouter.get('/getfriends',async(req,res)=>{
 
     const from= jwt.verify(token, "THISISMYSECRETKEY");
 
-    console.log(from.userid)
+    
 
     let allRequest = await Friend.find({user2:from.userid,areFriends: true});
-    console.log('allrequest',allRequest)
+   
     
         
     let allRequest2 = await Friend.find({user1:from.userid,areFriends: true});
-    console.log('allrequest2',allRequest2)
+    
 
     await Promise.all(allRequest2.map(async (elem) => {
         allRequest.push({...elem._doc,flag:true});
     }))
 
-    console.log(allRequest)
+    
 
     if(allRequest.length>0){
         const users=[];
@@ -169,20 +162,25 @@ friendsRouter.get('/getfriends',async(req,res)=>{
             }
         }));
 
-        console.log('users',users)
+        
 
         res.json({users})
     }
     else res.json({users:[]});
 })
 
-friendsRouter.get('/accept',async(req,res)=>{
+ friendsRouter.get('/accept',async(req,res)=>{
 
     const {token} = req.headers;
 
+    
     const from= jwt.verify(token, "THISISMYSECRETKEY")
 
+    console.log(from.userid, req.headers.id)
+
     const mid=await Friend.find({user2:from.userid,user1:req.headers.id, request:true});
+
+    
 
     if(mid.length==0)
     {
@@ -209,13 +207,13 @@ friendsRouter.get('/reject',async(req,res)=>{
 
     const from= jwt.verify(token, "THISISMYSECRETKEY");
 
-    console.log('REJECT',from.userid, req.headers.id)
+    
 
-    if(req.headers.friend==='true')
-   {console.log('hereeeeeeeee') 
-    const allRequest2 = await Friend.deleteOne({$or:[{user2:from.userid,user1:req.headers.id, areFriends:true},{user1:from.userid,user2:req.headers.id, areFriends:true}]});}
-   else{
-    const allRequest = await Friend.deleteOne({user1:from.userid,user2:req.headers.id, request:true});}
+    
+    const allRequest2 = await Friend.deleteOne({$or:[{user2:from.userid,user1:req.headers.id, areFriends:true},{user1:from.userid,user2:req.headers.id, areFriends:true}]});
+  
+    
+    const allRequest = await Friend.deleteOne({$or:[{user2:from.userid,user1:req.headers.id, request:true},{user1:from.userid,user2:req.headers.id, request:true}]});
 
     res.json({status:'successful'})
